// generated/models/CoflnetSongVoterModelsPartyPlaylistEntry.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Songvoter
 * Songvoter
 *
 * The version of the OpenAPI document: 0.0.1
 * Contact: support@coflnet.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { CoflnetSongVoterModelsPartyPlaylistEntrySelfVoteState } from './CoflnetSongVoterModelsPartyPlaylistEntrySelfVoteState';
import {
    CoflnetSongVoterModelsPartyPlaylistEntrySelfVoteStateFromJSON,
    CoflnetSongVoterModelsPartyPlaylistEntrySelfVoteStateFromJSONTyped,
    CoflnetSongVoterModelsPartyPlaylistEntrySelfVoteStateToJSON,
} from './CoflnetSongVoterModelsPartyPlaylistEntrySelfVoteState';
import type { CoflnetSongVoterModelsSong } from './CoflnetSongVoterModelsSong';
import {
    CoflnetSongVoterModelsSongFromJSON,
    CoflnetSongVoterModelsSongFromJSONTyped,
    CoflnetSongVoterModelsSongToJSON,
} from './CoflnetSongVoterModelsSong';

/**
 * 
 * @export
 * @interface CoflnetSongVoterModelsPartyPlaylistEntry
 */
export interface CoflnetSongVoterModelsPartyPlaylistEntry {
    /**
     * 
     * @type {CoflnetSongVoterModelsPartyPlaylistEntrySelfVoteState}
     * @memberof CoflnetSongVoterModelsPartyPlaylistEntry
     */
    selfVote?: CoflnetSongVoterModelsPartyPlaylistEntrySelfVoteState;
    /**
     * 
     * @type {CoflnetSongVoterModelsSong}
     * @memberof CoflnetSongVoterModelsPartyPlaylistEntry
     */
    song?: CoflnetSongVoterModelsSong;
    /**
     * 
     * @type {number}
     * @memberof CoflnetSongVoterModelsPartyPlaylistEntry
     */
    upVotes?: number;
    /**
     * 
     * @type {number}
     * @memberof CoflnetSongVoterModelsPartyPlaylistEntry
     */
    downVotes?: number;
}

/**
 * Check if a given object implements the CoflnetSongVoterModelsPartyPlaylistEntry interface.
 */
export function instanceOfCoflnetSongVoterModelsPartyPlaylistEntry(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function CoflnetSongVoterModelsPartyPlaylistEntryFromJSON(json: any): CoflnetSongVoterModelsPartyPlaylistEntry {
    return CoflnetSongVoterModelsPartyPlaylistEntryFromJSONTyped(json, false);
}

export function CoflnetSongVoterModelsPartyPlaylistEntryFromJSONTyped(json: any, ignoreDiscriminator: boolean): CoflnetSongVoterModelsPartyPlaylistEntry {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'selfVote': !exists(json, 'selfVote') ? undefined : CoflnetSongVoterModelsPartyPlaylistEntrySelfVoteStateFromJSON(json['selfVote']),
        'song': !exists(json, 'song') ? undefined : CoflnetSongVoterModelsSongFromJSON(json['song']),
        'upVotes': !exists(json, 'upVotes') ? undefined : json['upVotes'],
        'downVotes': !exists(json, 'downVotes') ? undefined : json['downVotes'],
    };
}

export function CoflnetSongVoterModelsPartyPlaylistEntryToJSON(value?: CoflnetSongVoterModelsPartyPlaylistEntry | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'selfVote': CoflnetSongVoterModelsPartyPlaylistEntrySelfVoteStateToJSON(value.selfVote),
        'song': CoflnetSongVoterModelsSongToJSON(value.song),
        'upVotes': value.upVotes,
        'downVotes': value.downVotes,
    };
}