// generated/models/CoflnetSongVoterModelsSong.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Songvoter
 * Songvoter
 *
 * The version of the OpenAPI document: 0.0.1
 * Contact: support@coflnet.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { CoflnetSongVoterModelsExternalSong } from './CoflnetSongVoterModelsExternalSong';
import {
    CoflnetSongVoterModelsExternalSongFromJSON,
    CoflnetSongVoterModelsExternalSongFromJSONTyped,
    CoflnetSongVoterModelsExternalSongToJSON,
} from './CoflnetSongVoterModelsExternalSong';

/**
 * 
 * @export
 * @interface CoflnetSongVoterModelsSong
 */
export interface CoflnetSongVoterModelsSong {
    /**
     * Gets or Sets Id
     * @type {string}
     * @memberof CoflnetSongVoterModelsSong
     */
    id?: string | null;
    /**
     * Gets or Sets Title
     * @type {string}
     * @memberof CoflnetSongVoterModelsSong
     */
    title?: string | null;
    /**
     * Gets or Sets Occurences
     * @type {Array<CoflnetSongVoterModelsExternalSong>}
     * @memberof CoflnetSongVoterModelsSong
     */
    occurences?: Array<CoflnetSongVoterModelsExternalSong> | null;
}

/**
 * Check if a given object implements the CoflnetSongVoterModelsSong interface.
 */
export function instanceOfCoflnetSongVoterModelsSong(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function CoflnetSongVoterModelsSongFromJSON(json: any): CoflnetSongVoterModelsSong {
    return CoflnetSongVoterModelsSongFromJSONTyped(json, false);
}

export function CoflnetSongVoterModelsSongFromJSONTyped(json: any, ignoreDiscriminator: boolean): CoflnetSongVoterModelsSong {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': !exists(json, 'id') ? undefined : json['id'],
        'title': !exists(json, 'title') ? undefined : json['title'],
        'occurences': !exists(json, 'occurences') ? undefined : (json['occurences'] === null ? null : (json['occurences'] as Array<any>).map(CoflnetSongVoterModelsExternalSongFromJSON)),
    };
}

export function CoflnetSongVoterModelsSongToJSON(value?: CoflnetSongVoterModelsSong | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'title': value.title,
        'occurences': value.occurences === undefined ? undefined : (value.occurences === null ? null : (value.occurences as Array<any>).map(CoflnetSongVoterModelsExternalSongToJSON)),
    };
}